// src/types/Enums/Descriptions/Rampart/DruidsDescription.ts
export enum DruidsDescription {
  AERIS = 'Аэрис – один из самых лучших знатоков животных в АвЛи. Он никогда не ошибается, когда объединяет в пары эльфов и пегасов, и это умение сделало его одним из самых незаменимых друидов в здешних местах.',
  ALAGAR = 'Говорят, что когда-то Алагар жил среди снежных эльфов в стране Вори, и потому его способности управлению водой, особенно замёрзшей, просто безграничны. Он начал служить АвЛи ещё до вторжения криган.',
  CORONIUS = 'Короний успел отучиться в одном из университетов Эрафии один семестр, прежде чем понял, что важность теоретических знаний переоценивают. Он отправился в АвЛи, где нашёл друида, который стал учить его на наглядных примерах.',
  ELESHAR = 'Эльшар по праву считается в АвЛи самым способным студентом из тех, кто обучается магии. Если его способности не угаснут в ближайшее несколько десятков лет, он, вероятно, станет самым молодым друидом, принятым в Совет Старейшин.',
  GEM = 'Джем была одной из величайших волшебниц, когда-либо ступавших по земле Энрота. Во время Войны за Престол она служила королю Роланду Айронфисту. Когда Роланд прочно обосновался на троне Энрота, Джем перебралась в Эрафию и поселилась в АвЛи.',
  MALCOM = 'Малком – один из немногих гномов, обладающих природным даром чародейства. Зачастую он может выучить заклинание, увидев лишь раз, как это делают другие. Его быстро призвали на службу и поручили защищать АвЛи от дьяволов Эофола.',
  MELODIA = 'Быть может, Мелодия – не самая сильная жрица друидов в АвЛи, но зато, бесспорно, самая удачливая. Ей удавалось одерживать громкие победы, даже когда все обстоятельства были против неё. Солдаты с великой радостью служат под её началом.',
  ULAND = 'Уланд много лет лечил раненых, прежде чем стать друидом, но знания, полученные им за время службы в армии, помогли ему заслужить репутацию отличного командира.'
}